moved utils function to separate modules
