+++ datacapsule_crossref/download_works.py
@@ -13,10 +13,9 @@ from six.moves.urllib.parse import urlencode
 from requests_futures.sessions import FuturesSession
 from tqdm import tqdm
 
-from datacapsule_crossref.utils import (
-  makedirs,
-  configure_session_retry
-)
+from datacapsule_crossref.utils.io import makedirs
+from datacapsule_crossref.utils.requests import configure_session_retry
+
 
 DEFLATE = "deflate"
 BZIP2 = "bzip2"
