+++ datacapsule_crossref/extract_summaries_from_works.py
@@ -13,10 +13,8 @@ from tqdm import tqdm
 from six import string_types
 from future.utils import raise_from
 
-from datacapsule_crossref.utils import (
-  makedirs,
-  write_csv
-)
+from datacapsule_crossref.utils.io import makedirs
+from datacapsule_crossref.utils.csv import write_csv
 
 from datacapsule_crossref.collection_utils import (
   iter_sort_window
