+++ datacapsule_crossref/extract_summaries_from_works_test.py
@@ -1,0 +1,7 @@
+from .extract_summaries_from_works import (
+  clean_text
+)
+
+class TestCleanText(object):
+  def test_(self):
+    pass
