+++ datacapsule_crossref/utils/collection.py
@@ -1,0 +1,5 @@
+def iter_dict_to_list(iterable, fields):
+  return (
+    [item.get(field) for field in fields]
+    for item in iterable
+  )
